zstream: comparison of integer expressions of different signedness
zstream_redup.c: In function 'spread':
zstream_redup.c:109:24: error: comparison of integer expressions of different signedness: 'ssize_t' {aka 'long int'} and 'size_t' {aka 'long unsigned int'} [-Werror=sign-compare]
  109 |         } else if (err != count) {
      |                        ^~
cc1: all warnings being treated as errors

zstream_dump.c: In function 'sprintf_bytes':
zstream_dump.c:154:23: error: comparison of integer expressions of different signedness: 'int' and 'uint_t' {aka 'unsigned int'} [-Werror=sign-compare]
  154 |         for (i = 0; i < buf_len; i++) {
      |                       ^
zstream_dump.c: In function 'zstream_do_dump':
zstream_dump.c:338:40: error: comparison of integer expressions of different signedness: 'int' and 'long long unsigned int' [-Werror=sign-compare]
  338 |                                 if (sz > SPA_MAXBLOCKSIZE) {
      |                                        ^
cc1: all warnings being treated as errors

+++ usr/src/cmd/zstream/zstream_redup.c
@@ -106,7 +106,7 @@ spread(int fd, void *buf, size_t count, off_t offset)
 		    "Error while reading file: %s\n",
 		    strerror(errno));
 		exit(1);
-	} else if (err != count) {
+	} else if ((size_t)err != count) {
 		(void) fprintf(stderr,
 		    "Error while reading file: short read\n");
 		exit(1);
